+++ src/components/slide/content/feed-selector.jsx
@@ -134,7 +134,9 @@ function FeedSelector({
           value={getValueFromConfiguration(element.name)}
           label={element.label}
           optionsEndpoint={element.endpoint}
-          singleSelect={formElement.singleSelect ?? false}
+          disableSearch={element.disableSearch ?? false}
+          singleSelect={element.singleSelect ?? false}
+          helpText={element.helpText ?? null}
         />
       );
     }