+++ CHANGELOG.md
@@ -9,6 +9,9 @@ All notable changes to this project will be documented in this file.
   - Fix save screen and go to list
   - Disable eslint for proptypes
   - Disable eslint for react in scope
+- [#284](https://github.com/os2display/display-admin-client/pull/284)
+  - Fixed spacer calculations for previews.
+  - Added border to preview container
 - [#283](https://github.com/os2display/display-admin-client/pull/283)
   - Fixed preview issues. 
 - [#275](https://github.com/os2display/display-admin-client/pull/275)